requirements.txt
--index-url https://download.pytorch.org/whl/nightly/cu132
torch
torchvision
torchaudio
Pillow
numpy
matplotlib
seaborn
scikit-learn